# mp-app/openapi.yaml
components:
  schemas:
...
    EquipmentCreate:
      additionalProperties: false
      properties:
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
        count:
          default: 1
          minimum: 1.0
          title: Count
          type: integer
        name:
          maxLength: 120
          minLength: 1
          title: Name
          type: string
      required:
      - name
      title: EquipmentCreate
      type: object
    GenerateRequest:
      additionalProperties: false
      properties:
        template:
          default: fb
          pattern: ^(ppl|ul|fb|corrective)$
          title: Template
          type: string
      title: GenerateRequest
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
info:
  description: Local-first Gym OS core API (offline-capable).
  title: Muscle Paradise Core
  version: 0.20.0
  x-generated-by: app.export_openapi
openapi: 3.1.0
paths:
...
  /api/v1/equipment:
    get:
      operationId: list_all_api_v1_equipment_get
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                title: Response List All Api V1 Equipment Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: All equipment
      tags:
      - equipment
    post:
      operationId: create_api_v1_equipment_post
      parameters:
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EquipmentCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Create Api V1 Equipment Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add equipment
      tags:
      - equipment
...
  /api/v1/members/{member_id}/programs:
    get:
      operationId: list_programs_api_v1_members__member_id__programs_get
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                title: Response List Programs Api V1 Members  Member Id  Programs Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Program history for a member
      tags:
      - programs
  /api/v1/members/{member_id}/programs/generate:
    post:
      operationId: generate_api_v1_members__member_id__programs_generate_post
      parameters:
      - in: path
        name: member_id
        required: true
        schema:
          title: Member Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Generate Api V1 Members  Member Id  Programs Generate Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Generate a rule-based program (draft)
      tags:
      - programs
...
  /api/v1/programs/{program_id}/apply:
    post:
      operationId: apply_api_v1_programs__program_id__apply_post
      parameters:
      - in: path
        name: program_id
        required: true
        schema:
          title: Program Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Apply Api V1 Programs  Program Id  Apply Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 'Approve + apply (C8: after dry-run)'
      tags:
      - programs
  /api/v1/programs/{program_id}/archive:
    post:
      operationId: archive_api_v1_programs__program_id__archive_post
      parameters:
      - in: path
        name: program_id
        required: true
        schema:
          title: Program Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Archive Api V1 Programs  Program Id  Archive Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Archive (the undo of apply)
      tags:
      - programs
  /api/v1/programs/{program_id}/dry-run:
    post:
      description: Re-validate the stored ops against the member's CURRENT filters.
      operationId: dry_run_api_v1_programs__program_id__dry_run_post
      parameters:
      - in: path
        name: program_id
        required: true
        schema:
          title: Program Id
          type: integer
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                additionalProperties: true
                title: Response Dry Run Api V1 Programs  Program Id  Dry Run Post
                type: object
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: 'Dry-run: validate without applying (C8)'
      tags:
      - programs
...